pyproject.toml
[tool.poetry]
name = "streamlit-analytics"
version = "0.1.0"
description = ""
authors = ["Matthew Tucker <matt.d.tucker@gmail.com>"]
readme = "README.md"
packages = [{include = "streamlit_analytics"}]

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.26.0"
openai = "^0.28.0"
langchain = "^0.0.293"
python-dotenv = "^1.0.0"
watchdog = "^3.0.0"
tiktoken = "^0.5.1"
llmonitor = "^0.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
